requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
anyio==4.3.0
apify_client==1.6.4
apify_shared==1.1.1
arxiv==2.1.0
asgiref==3.7.2
attrs==23.2.0
Automat==22.10.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
build==1.1.1
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
coloredlogs==15.0.1
constantly==23.10.4
cryptography==42.0.5
cssselect==1.2.0
dataclasses-json==0.5.14
Deprecated==1.2.14
distro==1.9.0
faiss-cpu==1.8.0
fastapi==0.110.0
feedparser==6.0.10
filelock==3.13.1
firebase-admin==6.5.0
Flask==3.0.2
Flask-Cors==4.0.0
flatbuffers==24.3.7
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.42
google==3.0.0
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-api-python-client==2.122.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.15.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-generativeai==0.4.1
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
googletrans==3.0.0
greenlet==3.0.3
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.3.1
httpcore==1.0.4
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.21.4
humanfriendly==10.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
importlib-metadata==6.11.0
importlib_resources==6.3.0
incremental==22.10.0
install==1.3.5
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kubernetes==29.0.0
langchain==0.0.316
langchain-community==0.0.29
langchain-core==0.1.33
langchain-google-genai==0.0.11
langchain-openai==0.1.1
langchain-pinecone==0.0.3
langchain-text-splitters==0.0.1
langchainhub==0.1.15
langsmith==0.1.31
llama-index==0.9.34
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numexpr==2.9.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnxruntime==1.17.1
openai==0.28.1
openapi-schema-pydantic==1.2.4
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
orjson==3.9.15
overrides==7.7.0
packaging==23.2
pandas==2.2.1
parsel==1.8.1
pillow==10.2.0
pinecone-client==3.0.0
playwright==1.42.0
posthog==3.5.0
Protego==0.3.0
proto-plus==1.23.0
protobuf==4.25.3
pulsar-client==3.4.0
pyarrow==15.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.14
pydeck==0.8.1b0
PyDispatcher==2.0.7
pyee==11.0.1
Pygments==2.17.2
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
pypdf==4.1.0
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
queuelib==1.6.2
referencing==0.34.0
regex==2023.12.25
reportlab==4.1.0
requests==2.31.0
requests-file==2.0.0
requests-oauthlib==1.4.0
rfc3986==1.5.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
Scrapy==2.11.1
sentence-transformers==2.6.0
service-identity==24.1.0
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.28
starlette==0.36.3
streamlit==1.32.2
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.4.0
tiktoken==0.6.0
tldextract==5.1.1
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.2.1
tornado==6.4
tqdm==4.66.2
transformers==4.39.1
triton==2.2.0
Twisted==24.3.0
typer==0.9.0
types-requests==2.31.0.20240311
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.28.0
uvloop==0.19.0
w3lib==2.1.2
watchdog==4.0.0
watchfiles==0.21.0
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.1
wrapt==1.16.0
yarl==1.9.4
zipp==3.18.0
zope.interface==6.2
